+++ CHANGES.md
@@ -1,6 +1,7 @@
 # Version 2.0.1 (to be released)
 
 * Added ToStringTransformer in core
+* Fixed compilation warning when adding rule input transformer using the GeneralValidatorBuilder
 
 # Version 2.0.0